requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
Pillow>=10.0.0
pytesseract>=0.3.10
python-slugify>=8.0.0
pandas>=2.0.0
tenacity>=8.2.0
tqdm>=4.66.0

